Clean up file structure
